Split ITF helper into separate file
